+++ app/package.json
@@ -22,6 +22,7 @@
     "@mdi/font": "^7.4.47",
     "@vueuse/core": "^14.2.1",
     "axios": "^1.13.1",
+    "github-markdown-css": "^5.9.0",
     "highlight.js": "^11.11.1",
     "markdown-it": "^14.1.1",
     "pinia": "^3.0.4",
